Add a catalog manager interface
